==> OcModels/Src/brooke_notebooks/03-blh-process-yellowtail-MOM6.ipynb <==
## Purpose

Filter and calculate annual and monthly means of monthly mixed layer depth, temperature, X and Y seawater velocities from MOM6 for yellowtail.

* For the annual means, each variable is filtered for latitudes, longitudes (based on bathymetry), and months that are relevant to that life stage.
* For the monthly means, each variables if filtered for latitudes and longitudes (based on bathymetry) that are relevant to a particular life stage. Means are calculated for every month.

## Set up

import xarray as xr
import numpy as np
import pandas as pd
import time

## Download data

Script assumes you have these files in a `OcModels/Data/MOM6` folder. To download necessary inputs, run:

*Depth (2-d)*
* *deptho* `wget -O ocean_static.deptho.nc https://psl.noaa.gov/thredds/fileServer/Projects/CEFI/regional_mom6/cefi_derivative/northeast_pacific/full_domain/hindcast/monthly/regrid/r20250509/static/ocean_static.deptho.nc`

*Mixed layer depth (2-d)*
* *MLD_003* `wget -O MLD_003.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc https://psl.noaa.gov/thredds/fileServer/Projects/CEFI/regional_mom6/cefi_portal/northeast_pacific/full_domain/hindcast/monthly/regrid/r20250509/MLD_003.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc`

*Temperature (3-d)*
* *thetao* `wget -O thetao.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc https://psl.noaa.gov/thredds/fileServer/Projects/CEFI/regional_mom6/cefi_portal/northeast_pacific/full_domain/hindcast/monthly/regrid/r20250509/thetao.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc`

*X & Y sea water velocities (3-d)*
* *uo_rotate* `wget -O uo_rotate.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc https://psl.noaa.gov/thredds/fileServer/Projects/CEFI/regional_mom6/cefi_portal/northeast_pacific/full_domain/hindcast/monthly/regrid/r20250509/uo_rotate.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc`
* *vo_rotate* `wget -O vo_rotate.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc https://psl.noaa.gov/thredds/fileServer/Projects/CEFI/regional_mom6/cefi_portal/northeast_pacific/full_domain/hindcast/monthly/regrid/r20250509/vo_rotate.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc`

Depth was based on a link Chia-Wei shared with me. The remaining queries were generated on the [CEFI Data Portal](https://psl.noaa.gov/cefi_portal/#data_access).
* Region: NEP
* Subregion: full domain
* Experiment Type: hindcast
* Output Frequency: monthly regrid
* Release: r20250509
* Data Category: ocean_monthly_z (for 3-d) or ocean_monthly (for 2-d)

In the future, it'd be nice to do this on the raw data rather than the gridded data.

# Annual Means

## Life stage dictionary

Based on [Data Availability](https://docs.google.com/document/d/1P8D0kH2xn4NYBc0ib3rYfSO0KAiyDNQ_kZgG4Qub7qY/edit?tab=t.0#heading=h.z5x77oqxpj4i), define filters for each lifestage for time, bottom depth, and latitude.

For northern yellowtail, the bottom depth is used to determine the longitudinal extent, and the average is taken across the whole water column. For example, if the longitudinal extent is bottom depths 90-180m, the netCDF is filtered for isobaths where the depth is \[90, 180], and the average is calculated for all the water \[0, 180] at those depths.

# assign months, latitudes, and depths to filter
# all are integers, depths are in meters
# filters are inclusive (closed interval), e.g. depth 0-180 includes depths 0 and 180 and every depth between

lifestage_dict = {
    "cop": { # copulation
        "min_month": 8,
        "max_month": 10,
        "min_lat": 40,
        "max_lat": 48,
        "min_depth": 90,
        "max_depth": 180
    }, 
    "part": { # partuition
        "min_month": 1,
        "max_month": 4,
        "min_lat": 40,
        "max_lat": 48,
        "min_depth": 0,
        "max_depth": 180
    },
    "larv": { # larvae
        "min_month": 2,
        "max_month": 3,
        "min_lat": 40,
        "max_lat": 48,
        "min_depth": 0,
        "max_depth": 90
    },
    "pjuv": { # pelagic juvenile
        "min_month": 4,
        "max_month": 8,
        "min_lat": 40,
        "max_lat": 48,
        "min_depth": 30,
        "max_depth": 130
    }
}

## Load datasets

# depth is used to filter longitudinal extent
ds_depth = xr.open_dataset("../../Data/MOM6/input/ocean_static.deptho.nc")

# mixed layer depth
ds_mld = xr.open_dataset("../../Data/MOM6/input/MLD_003.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc")

# temperature
ds_temp = xr.open_dataset("../../Data/MOM6/input/thetao.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc")

# X and Y sea water velocities
ds_xv = xr.open_dataset("../../Data/MOM6/input/uo_rotate.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc")
ds_yv = xr.open_dataset("../../Data/MOM6/input/vo_rotate.nep.full.hcast.monthly.regrid.r20250509.199301-202412.nc")

## Functions

'''
Print the size and dimensions of a dataset for a variable named 'var_name'
If nan_bool is true, also count the number of non-NA records (# and %) for that variable

Inputs:
- dataset: xarray Dataset with the variable var_name
- var_name: string with the variable name to count
- nan_bool: boolean, to print record count and % for the variable

Outputs:
- None, will print output
'''

def print_ds_info(dataset, var_name, nan_bool = False):
    denominator = dataset[var_name].size
    print(f'{denominator} records with dimensions {dataset[var_name].shape}')
    
    if nan_bool:
        numerator = np.isfinite(dataset[var_name].data).sum()
        percent = numerator / denominator * 100
        print(f'{numerator} records ({percent:.{2}f}%) are not np.nan')
        
    return None

In the future, it'd be nice to adjust these functions to have an argument to only run if the output file isn't in the directory already, so the notebook could run top to bototm and only re-caculate missing indices.

'''
Calculate annual mean for a 2-D variable

Inputs:
- var_dataset: xarray Dataset with the variable to average
- depth_dataset: xarray Dataset with depth to filter
- ds_var_name: string with the variable name, used to extract from var_dataset
- df_var_name: string with the variable name, used to rename column in returned dataframe
- filter_dict: dictionary with relevant filters
    - min_mon, max_mon: minimum and maximum month (integers, inclusive)
    - min_lat, max_lat: minimum and maximum latitude (integers or floats, inclusive)
    - min_depth, max_depth: minimum and maximum depth to filter the longitudinal extent
- print_counts: boolean, to print record count and % at each step; defaults to True for 2-D

Outputs:
- ds_df: pandas dataframe with an annual average of the variable
'''

def annual_mean_2d(var_dataset, depth_dataset, ds_var_name, df_var_name, filter_dict, print_counts = True):

    # start timer
    start_time = time.time()
    
    print("---" + df_var_name + "---")
    print_ds_info(var_dataset, ds_var_name, print_counts)
    
    # merge files
    ds_merged = xr.merge([var_dataset, depth_dataset])
    print_ds_info(ds_merged, ds_var_name, print_counts)

    # create list with months
    month_list = [x for x in range(filter_dict["min_month"], filter_dict["max_month"]+1)]
    # select times and latitudes
    ds_selected = ds_merged.sel(time =  ds_merged.time.dt.month.isin(month_list),
                                lat = slice(filter_dict["min_lat"], filter_dict["max_lat"]))
    print_ds_info(ds_selected, ds_var_name, print_counts)
    
    # mask depths
    ds_masked = ds_selected.where(
        (ds_selected.deptho >= filter_dict["min_depth"]) & 
        (ds_selected.deptho <= filter_dict["max_depth"])
    )
    print_ds_info(ds_masked, ds_var_name, print_counts)
    
    # calculate mean
    ds_mean = ds_masked[ds_var_name].groupby('time.year').mean(dim = ['time', 'lon', 'lat'])
    
    # create pandas dataframe
    ds_df = ds_mean.to_dataframe()
    
    # rename column from dataset to dataframe variable name
    ds_df.rename(columns = {ds_var_name:df_var_name}, inplace = True)

    # end timer and print runtime
    end_time = time.time()
    print(f'Runtime: {round((end_time - start_time) / 60, 2)} minutes')

    # write output
    ds_df.to_csv("../../Data/MOM6/interim/yellowtail_MOM6_" + df_var_name + ".csv")
    
    return ds_df

'''
Calculate annual mean for a 3-D variable

Inputs:
- var_dataset: xarray Dataset with the variable to average
- depth_dataset: xarray Dataset with depth to filter
- ds_var_name: string with the variable name, used to extract from var_dataset
- df_var_name: string with the variable name, used to rename column in returned dataframe
- filter_dict: dictionary with relevant filters
    - min_mon, max_mon: minimum and maximum month (integers, inclusive)
    - min_lat, max_lat: minimum and maximum latitude (integers or floats, inclusive)
    - min_depth, max_depth: minimum and maximum depth to filter the longitudinal extent
- print_counts: boolean, to print record count and % at each step; defaults to False for 3-D

Outputs:
- ds_df: pandas dataframe with an annual average of the variable
'''

def annual_mean_3d(var_dataset, depth_dataset, ds_var_name, df_var_name, filter_dict, print_counts = False):

    # start timer
    start_time = time.time()
    
    print("---" + df_var_name + "---")
    print_ds_info(var_dataset, ds_var_name, print_counts)
    
    # merge files
    ds_merged = xr.merge([var_dataset, depth_dataset])
    print_ds_info(ds_merged, ds_var_name, print_counts)

    # create list with months
    month_list = [x for x in range(filter_dict["min_month"], filter_dict["max_month"]+1)]
    # select times and latitudes
    ds_selected = ds_merged.sel(time =  ds_merged.time.dt.month.isin(month_list),
                                lat = slice(filter_dict["min_lat"], filter_dict["max_lat"]))
    print_ds_info(ds_selected, ds_var_name, print_counts)
    
    # mask depths
    ds_masked = ds_selected.where(
        (ds_selected.deptho >= filter_dict["min_depth"]) & 
        (ds_selected.deptho <= filter_dict["max_depth"])
    )
    print_ds_info(ds_masked, ds_var_name, print_counts)
    
    # calculate mean
    ds_mean = ds_masked[ds_var_name].groupby('time.year').mean(dim = ['time', 'z_l', 'lon', 'lat'])
    
    # create pandas dataframe
    ds_df = ds_mean.to_dataframe()
    
    # rename column from dataset to dataframe variable name
    ds_df.rename(columns = {ds_var_name:df_var_name}, inplace = True)

    # end timer and print runtime
    end_time = time.time()
    print(f'Runtime: {round((end_time - start_time) / 60, 2)} minutes')

    # write output
    ds_df.to_csv("../../Data/MOM6/interim/yellowtail_MOM6_" + df_var_name + ".csv")
    
    return ds_df

## Calculate means

# mixed layer depths
MLDpart = annual_mean_2d(ds_mld, ds_depth, "MLD_003", "MLDpart", lifestage_dict['part'], True)
MLDlarv = annual_mean_2d(ds_mld, ds_depth, "MLD_003", "MLDlarv", lifestage_dict['larv'], True)
MLDpjuv = annual_mean_2d(ds_mld, ds_depth, "MLD_003", "MLDpjuv", lifestage_dict['pjuv'], True)

# temperatures
Tcop = annual_mean_3d(ds_temp, ds_depth, "thetao", "Tcop", lifestage_dict['cop'], False)
Tpart = annual_mean_3d(ds_temp, ds_depth, "thetao", "Tpart", lifestage_dict['part'], False)

# X sea water velocities
XVlarv = annual_mean_3d(ds_xv, ds_depth, "uo_rotate", "XVlarv", lifestage_dict['larv'], False)
XVpjuv = annual_mean_3d(ds_xv, ds_depth, "uo_rotate", "XVpjuv", lifestage_dict['pjuv'], False)

# Y sea water velocities
YVlarv = annual_mean_3d(ds_yv, ds_depth, "vo_rotate", "YVlarv", lifestage_dict['larv'], False)
YVpjuv = annual_mean_3d(ds_yv, ds_depth, "vo_rotate", "YVpjuv", lifestage_dict['pjuv'], False)

## Lag indices as needed

For life stages where data from the prior calendar year are relevant for this year's recruitment deviaitons, adjust the index +1. For yellowtail, those life stages are preconditioning, copulation, and egg fertilization.

In the future, it'd be nice to do this programatically by adding the lag to the lifestage dictionary.

# lag computation data by one year
Tcop.index += 1

## Combine and write output

# merge pandas dataframes
output_df = pd.concat([MLDpart, MLDlarv, MLDpjuv, Tcop, Tpart, XVlarv, XVpjuv, YVlarv, YVpjuv], axis = 1)

# sort indices
output_df.sort_index(inplace = True)

# save result
output_df.to_csv("../../Data/MOM6/yellowtail_MOM6.csv")

# preview result
output_df

# Monthly Means

This section was written to look at monthly patterns for the same variables, and to match Megan's format of monthly GLORYS data.

'''
Calculate monthly mean for a 2-D variable

Inputs:
- var_dataset: xarray Dataset with the variable to average
- depth_dataset: xarray Dataset with depth to filter
- ds_var_name: string with the variable name, used to extract from var_dataset
- df_var_name: string with the variable name, used to rename column in returned dataframe, will have depth range appended
- min_lat, max_lat: minimum and maximum latitude (integers or floats, inclusive)
- min_depth, max_depth: minimum and maximum depth to filter the longitudinal extent

Outputs:
- ds_df: pandas dataframe with a filtered monthly average of the variable
'''

def monthly_mean_2d(var_dataset, depth_dataset, ds_var_name, df_var_name, min_lat, max_lat, min_depth, max_depth):
    
    # start timer
    start_time = time.time()
    
    print("---" + df_var_name + " (" + str(min_depth) + "-" + str(max_depth) + ") ---")
    print_ds_info(var_dataset, ds_var_name, nan_bool = False)
    
    # merge files
    ds_merged = xr.merge([var_dataset, depth_dataset])
    print_ds_info(ds_merged, ds_var_name, nan_bool = False)
    
    # select latitudes
    ds_selected = ds_merged.sel(lat = slice(min_lat, max_lat))
    print_ds_info(ds_selected, ds_var_name, nan_bool = False)
    
    # mask depths
    ds_masked = ds_selected.where((ds_selected.deptho >= min_depth) & (ds_selected.deptho <= max_depth))
    print_ds_info(ds_masked, ds_var_name, nan_bool = False)
    
    # calculate mean
    ds_mean = ds_masked[ds_var_name].groupby(['time.year', 'time.month']).mean(dim = ['time', 'lon', 'lat'])
    
    # create pandas dataframe
    ds_df = ds_mean.to_dataframe()
    
    # rename column from dataset to dataframe variable name
    df_var_name_with_depth = df_var_name + "_" + str(min_depth) + "_" + str(max_depth)
    ds_df.rename(columns = {ds_var_name:df_var_name_with_depth}, inplace = True)
    
    # end timer and print runtime
    end_time = time.time()
    print(f'Runtime: {round((end_time - start_time) / 60, 2)} minutes')
    
    # write output
    ds_df.to_csv("../../Data/MOM6/interim/yellowtail_MOM6_monthly_" + df_var_name_with_depth + ".csv")
    
    return ds_df

'''
Calculate monthly mean for a 3-D variable

Inputs:
- var_dataset: xarray Dataset with the variable to average
- depth_dataset: xarray Dataset with depth to filter
- ds_var_name: string with the variable name, used to extract from var_dataset
- df_var_name: string with the variable name, used to rename column in returned dataframe, will have depth range appended
- min_lat, max_lat: minimum and maximum latitude (integers or floats, inclusive)
- min_depth, max_depth: minimum and maximum depth to filter the longitudinal extent

Outputs:
- ds_df: pandas dataframe with a filtered monthly average of the variable
'''

def monthly_mean_3d(var_dataset, depth_dataset, ds_var_name, df_var_name, min_lat, max_lat, min_depth, max_depth):
    
    # start timer
    start_time = time.time()
    
    print("---" + df_var_name + " (" + str(min_depth) + "-" + str(max_depth) + ") ---")
    print_ds_info(var_dataset, ds_var_name, nan_bool = False)
    
    # merge files
    ds_merged = xr.merge([var_dataset, depth_dataset])
    print_ds_info(ds_merged, ds_var_name, nan_bool = False)
    
    # select latitudes
    ds_selected = ds_merged.sel(lat = slice(min_lat, max_lat))
    print_ds_info(ds_selected, ds_var_name, nan_bool = False)
    
    # mask depths
    ds_masked = ds_selected.where((ds_selected.deptho >= min_depth) & (ds_selected.deptho <= max_depth))
    print_ds_info(ds_masked, ds_var_name, nan_bool = False)
    
    # calculate mean
    ds_mean = ds_masked[ds_var_name].groupby(['time.year', 'time.month']).mean(dim = ['time', 'z_l', 'lon', 'lat'])
    
    # create pandas dataframe
    ds_df = ds_mean.to_dataframe()
    
    # rename column from dataset to dataframe variable name
    df_var_name_with_depth = df_var_name + "_" + str(min_depth) + "_" + str(max_depth)
    ds_df.rename(columns = {ds_var_name:df_var_name_with_depth}, inplace = True)
    
    # end timer and print runtime
    end_time = time.time()
    print(f'Runtime: {round((end_time - start_time) / 60, 2)} minutes')
    
    # write output
    ds_df.to_csv("../../Data/MOM6/interim/yellowtail_MOM6_monthly_" + df_var_name_with_depth + ".csv")
    
    return ds_df

It would be more concise code if the 2-d and 3-d functions were the same. The only real difference is the calculate mean step, when `'z_l'` is included as a dimension for a 3-d variable and it is not for a 2-d variable.

## Depth dictionary

Based on [Data Availability](https://docs.google.com/document/d/1P8D0kH2xn4NYBc0ib3rYfSO0KAiyDNQ_kZgG4Qub7qY/edit?tab=t.0#heading=h.z5x77oqxpj4i), define filters bottom depth and latitude.

For northern yellowtail, the bottom depth is used to determine the longitudinal extent, and the average is taken across the whole water column. For example, if the longitudinal extent is bottom depths 90-180m, the netCDF is filtered for isobaths where the depth is \[90, 180], and the average is calculated for all the water \[0, 180] at those depths.

# assign depths to filter, in meters
# filters are inclusive (closed interval), e.g. depth 0-180 includes depths 0 and 180 and every depth between

depth_dict = {
    "pre_cop_egg": { # preconditioning, copulation, and egg fertilization
        "min_depth": 90,
        "max_depth": 180
    }, 
    "part": { # partuition
        "min_depth": 0,
        "max_depth": 180
    },
    "larv": { # larvae
        "min_depth": 0,
        "max_depth": 90
    },
    "pjuv": { # pelagic juvenile
        "min_depth": 30,
        "max_depth": 130
    },
    "bjuv": { # benthic juvenile
        "min_depth": 180,
        "max_depth": 550
    }
}

## Calculate means

# mixed layer depths
MLD_list = [monthly_mean_2d(ds_mld, ds_depth, "MLD_003", "MLD", min_lat = 40, max_lat = 48, min_depth = v['min_depth'], max_depth = v['max_depth']) for v in depth_dict.values()]

# temperatures
T_list = [monthly_mean_3d(ds_temp, ds_depth, "thetao", "T", min_lat = 40, max_lat = 48, min_depth = v['min_depth'], max_depth = v['max_depth']) for v in depth_dict.values()]

# X sea water velocities
XV_list = [monthly_mean_3d(ds_xv, ds_depth, "uo_rotate", "XV", min_lat = 40, max_lat = 48, min_depth = v['min_depth'], max_depth = v['max_depth']) for v in depth_dict.values()]

# Y sea water velocities
YV_list = [monthly_mean_3d(ds_yv, ds_depth, "vo_rotate", "YV", min_lat = 40, max_lat = 48, min_depth = v['min_depth'], max_depth = v['max_depth']) for v in depth_dict.values()]

## Combine and write output

# merge pandas dataframes for each variable
MLD_monthly = pd.concat(MLD_list, axis = 1)
T_monthly = pd.concat(T_list, axis = 1)
XV_monthly = pd.concat(XV_list, axis = 1)
YV_monthly = pd.concat(YV_list, axis = 1)

# merge pandas dataframes for all variables together
output_monthly_df = pd.concat([MLD_monthly, T_monthly, XV_monthly, YV_monthly], axis = 1)

# save result
output_monthly_df.to_csv("../../Data/MOM6/yellowtail_MOM6_monthly.csv")

# preview result
output_monthly_df.head()
