==> OcModels/Src/brooke_notebooks/04-quick-check-monthly-means.ipynb <==
import pandas as pd

# load annual and monthly data
mom6_annual = pd.read_csv("../../Data/Yellowtail/yellowtail_MOM6.csv", index_col = "year")
mom6_monthly = pd.read_csv("../../Data/Yellowtail/yellowtail_MOM6_monthly.csv", index_col = "year")

# calculate annual mean from monthly data
mom6_monthly_annual = mom6_monthly.copy().groupby("year").mean().drop(columns = ['month'])

# calculate mean % difference for each variable (should be quite small)
((mom6_monthly_annual - mom6_annual) / mom6_annual * 100).mean()
