==> OcModels/Src/brooke_notebooks/05-remove-2024-mom6-data.ipynb <==
There was an issue with the 2024 data in the model run of MOM6 I pulled and post-processed data from. I am removing the 2024 onwards data from the two files retroactively, just so we don't accidentally use the data for comparisons in the future.

import pandas as pd

Load data.

mom6_annual = pd.read_csv("../../Data/Yellowtail/yellowtail_MOM6.csv", index_col = "year")
mom6_monthly = pd.read_csv("../../Data/Yellowtail/yellowtail_MOM6_monthly.csv", index_col = "year")

Check tail of dataframes, see that 2024 and 2025 data is included.

mom6_annual.tail()

mom6_monthly.tail()

Filter out 2024 data onwards.

(Note on `Tcop` - Copulation was lagged one year forward, since temperature for copulation one year impacts the next year's recruitment, i.e. 2023 MOM6 temperature data is used to calculate `Tcop` for 2024. This step filters out 2024 `Tcop` which is not impacted by the MOM6 data issue - so the filter is a little more conservative than needed. Since we haven't looked at comparisons for years that we don't have all the life stage data for anyways, this doesn't impact our comparisons.)

mom6_annual = mom6_annual.loc[mom6_annual.index < 2024]
mom6_monthly = mom6_monthly.loc[mom6_monthly.index < 2024]

Check tail of dataframes, verify that 2024 data is removed.

mom6_annual.tail()

mom6_monthly.tail()

Rewrite dataframes without 2024 data.

mom6_annual.to_csv("../../Data/Yellowtail/yellowtail_MOM6.csv")
mom6_monthly.to_csv("../../Data/Yellowtail/yellowtail_MOM6_monthly.csv")
